requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-docx>=0.8.11
Pillow>=10.0.0
lxml>=4.9.0
